load_all example
