+++ test/regressions/manifests/poddefault.yaml
@@ -8,12 +8,12 @@ spec:
       orbit/stuff: ""
   desc: Orbit Stuff
 ---
-# apiVersion: kubeflow.org/v1alpha1
-# kind: PodDefault
-# metadata:
-#   name: orbit-stuff-2
-# spec:
-#   selector:
-#     matchLabels:
-#       orbit/stuff: ""
-#   desc: Orbit Stuff
+apiVersion: kubeflow.org/v1alpha1
+kind: PodDefault
+metadata:
+  name: orbit-stuff-2
+spec:
+  selector:
+    matchLabels:
+      orbit/stuff: ""
+  desc: Orbit Stuff
